src/Azure/Functions/PythonAPI/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
pandas
psycopg2-binary
sqlalchemy
mysql-connector-python
# mysql
# mysql-connector
# mysql.connector
six
python-jose
pymysql
mysqlclient
